% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-Exception.R
\name{package-Exception}
\alias{package-Exception}
\title{Classed exception object hierarchy}
\description{
Provides classed exception objects with a pre-defined hierarchy of classes.
This allows catching exceptions in a nested way by using a class or parent
class name with try-catch.
}

